Moved content to src folder, added redux logger
