+++ src/api/registerForPushNotificationsAsync.js
