+++ src/store/index.js
@@ -1,8 +1,10 @@
 import { createStore, applyMiddleware } from 'redux'
 import { Reducers } from '../reducers'
 import thunk from 'redux-thunk'
+import logger from 'redux-logger'
 
 export const store = createStore(
   Reducers,
-  applyMiddleware(thunk)
+  applyMiddleware(thunk),
+  applyMiddleware(logger)
 )
